Serialize CryptoKey using CoreIPC https://bugs.webkit.org/show_bug.cgi?id=272064 rdar://128551948
Reviewed by Alex Christensen and David Kilzer.

Over the wire bytes will now be serialized using CoreIPC ::decode
and thus the crash in PropertyListDeserialization is avoided.
Note that serialization is still property list and that maintains backward
compatibility with serialized CryptoKeys on disks of existing users.

Testing: No new tests are added as this functionality is already being exercised by
Indexeddb tests/service worker tests.

* Source/WebCore/Headers.cmake:
* Source/WebCore/crypto/WrappedCryptoKey.h:
* Source/WebCore/crypto/SerializedCryptoKeyWrap.h:
* Source/WebCore/crypto/cocoa/SerializedCryptoKeyWrapMac.mm:
(WebCore::wrapSerializedCryptoKey):
(WebCore::createArrayFromData):
(WebCore::readSerializedCryptoKey):
(WebCore::unwrapCryptoKey):
(WebCore::unwrapSerializedCryptoKey): Deleted.
* Source/WebCore/crypto/gcrypt/SerializedCryptoKeyWrapGCrypt.cpp:
(WebCore::readSerializedCryptoKey):
(WebCore::unwrapCryptoKey):
(WebCore::unwrapSerializedCryptoKey): Deleted.
* Source/WebCore/crypto/openssl/SerializedCryptoKeyWrapOpenSSL.cpp:
(WebCore::readSerializedCryptoKey):
(WebCore::unwrapCryptoKey):
(WebCore::unwrapSerializedCryptoKey): Deleted.
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::unwrapCryptoKey):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.messages.in:
* Source/WebKit/UIProcess/WebProcessProxy.cpp:
(WebKit::WebProcessProxy::unwrapCryptoKey):
* Source/WebKit/UIProcess/WebProcessProxy.h:
* Source/WebKit/UIProcess/WebProcessProxy.messages.in:
* Source/WebKit/WebProcess/WebCoreSupport/WebCryptoClient.cpp:
(WebKit::WebCryptoClient::unwrapCryptoKey const):
* Source/WebKitLegacy/WebCoreSupport/WebCryptoClient.mm:
(WebCryptoClient::unwrapCryptoKey const):
* Tools/TestWebKitAPI/Tests/WebCore/cocoa/SerializedCryptoKeyWrap.mm:
(TestWebKitAPI::TEST_F):

Originally-landed-as: 272448.867@safari-7618-branch (b4edd8f7193f). rdar://128551948

test

Canonical link: https://commits.webkit.org/279933@main
