+++ Source/WebCore/crypto/openssl/SerializedCryptoKeyWrapOpenSSL.cpp
@@ -28,6 +28,7 @@
 
 #include "CryptoAlgorithmAESCTR.h"
 #include "NotImplemented.h"
+#include "WrappedCryptoKey.h"
 
 namespace WebCore {
 
@@ -51,13 +52,17 @@ bool wrapSerializedCryptoKey(const Vector<uint8_t>& masterKey, const Vector<uint
     return true;
 }
 
-bool unwrapSerializedCryptoKey(const Vector<uint8_t>& masterKey, const Vector<uint8_t>& wrappedKey, Vector<uint8_t>& key)
+std::optional<struct WrappedCryptoKey> readSerializedCryptoKey(const Vector<uint8_t>& wrappedKey)
 {
-    UNUSED_PARAM(masterKey);
+    std::array<uint8_t, 24> a { 0 };
+    std::array<uint8_t, 16> b { 0 };
+    struct WrappedCryptoKey k { a, wrappedKey, b };
+    return k;
+}
 
-    // No unwrapping performed -- the serialized key data is copied into the `key` variable.
-    key = Vector<uint8_t>(wrappedKey);
-    return true;
+std::optional<Vector<uint8_t>> unwrapCryptoKey([[maybe_unused]] const Vector<uint8_t>& masterKey, const struct WrappedCryptoKey& wrappedKey)
+{
+    return wrappedKey.encryptedKey;
 }
 
 } // namespace WebCore
