+++ Source/WebKit/WebProcess/WebCoreSupport/WebCryptoClient.cpp
@@ -30,6 +30,8 @@
 #include "WebPageProxyMessages.h"
 #include "WebProcess.h"
 #include "WebProcessProxyMessages.h"
+#include <WebCore/SerializedCryptoKeyWrap.h>
+#include <WebCore/WrappedCryptoKey.h>
 
 namespace WebKit {
 
@@ -48,13 +50,17 @@ std::optional<Vector<uint8_t>> WebCryptoClient::wrapCryptoKey(const Vector<uint8
 
 std::optional<Vector<uint8_t>> WebCryptoClient::unwrapCryptoKey(const Vector<uint8_t>& wrappedKey) const
 {
+    auto deserializedKey = WebCore::readSerializedCryptoKey(wrappedKey);
+    if (!deserializedKey)
+        return std::nullopt;
+
     if (m_pageIdentifier) {
-        auto sendResult = WebProcess::singleton().parentProcessConnection()->sendSync(Messages::WebPageProxy::UnwrapCryptoKey(wrappedKey), *m_pageIdentifier);
+        auto sendResult = WebProcess::singleton().parentProcessConnection()->sendSync(Messages::WebPageProxy::UnwrapCryptoKey(*deserializedKey), *m_pageIdentifier);
         auto [unwrappedKey] = sendResult.takeReplyOr(std::nullopt);
         return unwrappedKey;
     }
 
-    auto sendResult = WebProcess::singleton().parentProcessConnection()->sendSync(Messages::WebProcessProxy::UnwrapCryptoKey(wrappedKey), 0);
+    auto sendResult = WebProcess::singleton().parentProcessConnection()->sendSync(Messages::WebProcessProxy::UnwrapCryptoKey(*deserializedKey), 0);
     auto [unwrappedKey] = sendResult.takeReplyOr(std::nullopt);
     return unwrappedKey;
 }
